// src/app/core/api/generated/model/importAttendanceRawCommand.ts
/**
 * Clean Architecture WebAPI
 * An example of Clean Architecture WebAPI with Swagger and JWT Authentication
 *
 * The version of the OpenAPI document: v1
 * Contact: kamrantajerbashi@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AttendanceRawDto } from './attendanceRawDto';


export interface ImportAttendanceRawCommand { 
    punches?: Array<AttendanceRawDto> | null;
}